pyproject.toml
[tool.poetry]
name = "social-bots"
version = "0.1.0"
description = ""
authors = ["GladeOwl <owloftheglade@gmail.com>"]
readme = "README.md"
packages = [{include = "social_bots"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.4"
python-dotenv = "^1.0.0"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
requests = "^2.31.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
